fix: refactor AbortedError handling and reintroduce AbortedError class
