+++ packages/h5p-server/src/helpers/AbortedError.ts
